# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_tao_api_nhom_post": {
        "properties": {
          "ten": {
            "title": "Ten",
            "type": "string"
          }
        },
        "required": [
          "ten"
        ],
        "title": "Body_tao_api_nhom_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Dich phu de video Anh -> Viet",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/nhom": {
      "get": {
        "operationId": "danh_sach_api_nhom_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sap_xep",
            "required": false,
            "schema": {
              "default": "ten",
              "title": "Sap Xep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thu_tu",
            "required": false,
            "schema": {
              "default": "asc",
              "title": "Thu Tu",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "trang",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Trang",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "moi_trang",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Moi Trang",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Danh Sach Api Nhom Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh Sach",
        "tags": [
          "nhom"
        ]
      },
      "post": {
        "operationId": "tao_api_nhom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_tao_api_nhom_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tao Api Nhom Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tao",
        "tags": [
          "nhom"
        ]
      }
    },
    "/api/nhom/{ten}": {
      "delete": {
        "operationId": "xoa_api_nhom__ten__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Xoa",
        "tags": [
          "nhom"
        ]
      }
    },
    "/api/nhom/{ten}/hop": {
      "get": {
        "operationId": "doc_hop_api_nhom__ten__hop_get",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Doc Hop Api Nhom  Ten  Hop Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Doc Hop",
        "tags": [
          "nhom"
        ]
      },
      "post": {
        "description": "Hop di qua dung validator cua CLI; frontend kiem them chi de bao som.",
        "operationId": "dat_hop_api_nhom__ten__hop_post",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Than",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dat Hop Api Nhom  Ten  Hop Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dat Hop",
        "tags": [
          "nhom"
        ]
      }
    },
    "/api/nhom/{ten}/thuat-ngu": {
      "get": {
        "operationId": "doc_thuat_ngu_api_nhom__ten__thuat_ngu_get",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sap_xep",
            "required": false,
            "schema": {
              "default": "goc",
              "title": "Sap Xep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thu_tu",
            "required": false,
            "schema": {
              "default": "asc",
              "title": "Thu Tu",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "trang",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Trang",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "moi_trang",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Moi Trang",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Doc Thuat Ngu Api Nhom  Ten  Thuat Ngu Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Doc Thuat Ngu",
        "tags": [
          "nhom"
        ]
      },
      "post": {
        "operationId": "dat_thuat_ngu_api_nhom__ten__thuat_ngu_post",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Than",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Dat Thuat Ngu Api Nhom  Ten  Thuat Ngu Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dat Thuat Ngu",
        "tags": [
          "nhom"
        ]
      }
    },
    "/api/nhom/{ten}/thuat-ngu/{goc}": {
      "delete": {
        "operationId": "xoa_thuat_ngu_api_nhom__ten__thuat_ngu__goc__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ten",
            "required": true,
            "schema": {
              "title": "Ten",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "goc",
            "required": true,
            "schema": {
              "title": "Goc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Xoa Thuat Ngu",
        "tags": [
          "nhom"
        ]
      }
    },
...
  }
}